% Generated by roxygen2 (4.0.1): do not edit by hand
\name{areas}
\alias{areas}
\alias{areas.SpatialPolygons}
\alias{areas.SpatialPolygonsDataFrame}
\title{To calculate the area of each (projected) polygon in a SpatialPolygons* object}
\usage{
areas(spobj, CRSobj, ...)

\method{areas}{SpatialPolygons}(spobj, CRSobj, ...)

\method{areas}{SpatialPolygonsDataFrame}(spobj, CRSobj, ...)
}
\arguments{
\item{spobj}{SpatialPolygons* object}

\item{CRSobj}{for projection}

\item{\dots}{ignored}
}
\description{
To calculate the area of each (projected) polygon in a SpatialPolygons* object
}

