% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CA_regions}
\alias{CA_regions}
\title{Geographical boundaries and tract identifiers for major CA regions}
\format{A list with four items: "San Joaquin", "South Coast", "Bay Area", and "Other".
  Each element is a list with two items:
  \itemize{
    \item FIPS. A character vector of FIPS codes for tracts belonging to the region.
    \item boundary. A SpatialPolygons object, created by merging tract geometries.
  }}
\usage{
tract_regions
}
\description{
Based on data provided by the Bay Area AQMD
}

