% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalEnviroScreen-data.R
\docType{data}
\name{CA_tracts}
\alias{CA_tracts}
\title{California census tracts}
\format{A SpatialPolygonsDataFrame}
\source{
\url{http://oehha.ca.gov/ej/ces2.html}
}
\usage{
CA_tracts
}
\description{
California census tracts, with areas computed from an Albers projection
}

