% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{CalEnviroScreen}
\alias{CalEnviroScreen}
\alias{CalEnviroScreen-package}
\title{CalEnviroScreen}
\description{
This package provides CalEnviroScreen data in native R format.
}
\author{
David Holstius \email{david.holstius@berkeley.edu}
}
\references{
Meehan August L, Faust JB, Cushing L, Zeise L, Alexeeff, GV (2012). Methodological Considerations in Screening for Cumulative Environmental Health Impacts: Lessons Learned from a Pilot Study in California. Int J Environ Res Public Health 9(9): 3069-3084.
}
\keyword{environmental}
\keyword{model}
\keyword{package}
\keyword{screening}

