% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CES2.R
\name{CES2_rank}
\alias{CES2_pctl}
\alias{CES2_rank}
\alias{compute_CES2_scores}
\alias{compute_CES2_subscores}
\title{Calculate percentiles (as in CES 2.x)}
\usage{
CES2_rank(x, i = which(x > 0), ...)

CES2_pctl(x, ...)

compute_CES2_subscores(.data, min_obs, unit_var = "FIPS",
  group_var = "Group", ...)

compute_CES2_scores(.data, breaks = CES_PERCENTILE_BREAKS)
}
\arguments{
\item{x}{values}

\item{i}{(optional) expression indicating which values to include in the calculation}

\item{.data}{data frame}

\item{min_obs}{minimum number of observations in a group (else NA)}

\item{unit_var}{character}

\item{group_var}{character}

\item{breaks}{named numeric vector}

\item{\dots}{further arguments to base::rank}

\item{x}{values}

\item{\dots}{further arguments to \code{CES2_rank}}

\item{\dots}{ignored}

\item{.data}{data frame}
}
\description{
Calculate percentiles (as in CES 2.x)

Compute subscores (as in CES 2.x)

Compute scores (as in CES 2.x)
}

