% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CES2_data}
\alias{CES2_data}
\title{CalEnviroScreen 2.0 data}
\format{data.frame}
\source{
\url{http://oehha.ca.gov/ej/ces2.html}
}
\usage{
CES2_data
}
\description{
Published data for CalEnviroScreen 2.0, with percentiles replicated from inferred methods
}

