% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CES2_PERCENTILE_BREAKS}
\alias{CES2_PERCENTILE_BREAKS}
\title{Percentile breaks (inferred from published CES 2.0 data)}
\format{named numeric vector}
\usage{
data(CES2_metadata); CES2_PERCENTILE_BREAKS
}
\description{
Percentile breaks (inferred from published CES 2.0 data)
}

