% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CES2_POLLUTION_VARS}
\alias{CES2_POLLUTION_VARS}
\title{Variables associated with pollution burden}
\format{named character vector, with names corresponding to column names in the published Excel file}
\usage{
data(CES2_metadata); CES2_POLLUTION_VARS
}
\description{
Names of variables used to calculate pollution burden score
}

