% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalEnviroScreen-data.R
\docType{data}
\name{CES2_WEIGHTS}
\alias{CES2_WEIGHTS}
\title{Relative weights used to compute CES 2.0 scores}
\format{named numeric vector}
\usage{
data(CES2_metadata); CES2_WEIGHTS
}
\description{
Relative weights used to compute CES 2.0 scores
}

