% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CES3.R
\name{CES3_rank}
\alias{CES3_pctl}
\alias{CES3_rank}
\alias{compute_CES3_scores}
\alias{compute_CES3_subscores}
\title{Calculate percentiles (as in CES3)}
\usage{
CES3_rank(x, i = which(x > 0), ...)

CES3_pctl(x, ...)

compute_CES3_subscores(.data, min_obs, unit_var = "FIPS",
  group_var = "Group", ...)

compute_CES3_scores(.data, breaks = CES_PERCENTILE_BREAKS)
}
\arguments{
\item{x}{values}

\item{i}{(optional) expression indicating which values to include in the calculation}

\item{.data}{a dplyr::tbl}

\item{min_obs}{minimum number of observations in a group (else NA)}

\item{unit_var}{character}

\item{group_var}{character}

\item{breaks}{named numeric vector}

\item{\dots}{further arguments to base::rank}

\item{x}{values}

\item{\dots}{further arguments to \link{CES3_rank}}

\item{\dots}{ignored}

\item{.data}{data frame}
}
\description{
Calculate percentiles (as in CES3)

To compute CES3 subscores

Compute scores (as in CES 3.0)
}

