% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CES_rescale}
\alias{CES_rescale}
\title{To normalize a vector}
\usage{
CES_rescale(x, digits = Inf, lower = min(x, na.rm = TRUE))
}
\arguments{
\item{x}{numeric}

\item{digits}{numeric}

\item{lower}{numeric}

\item{na.rm}{logical}
}
\description{
To normalize a vector
}
\note{
upper is always 1.0
}

