% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comp_rank}
\alias{comp_rank}
\title{Calculate ranks (as in CalEnviroScreen 2.0)}
\usage{
comp_rank(x, i = which(x > 0), ...)
}
\arguments{
\item{x}{values}

\item{i}{(optional) expression indicating which values to include in the calculation}

\item{\dots}{further arguments to base::rank}
}
\description{
Calculate ranks (as in CalEnviroScreen 2.0)
}

