% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute_CES2_avg_pctls}
\alias{compute_CES2_avg_pctls}
\title{To average CES2 percentiles}
\usage{
compute_CES2_avg_pctls(.data, min_obs, ...)
}
\arguments{
\item{.data}{a dplyr::tbl}

\item{min_obs}{minimum number of observations in a group (else NA)}

\item{\dots}{ignored}
}
\description{
To average CES2 percentiles
}

