% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute_CES2_scores}
\alias{compute_CES2_scores}
\title{To compute CES2 scores}
\usage{
compute_CES2_scores(.data, breaks = CES2_PERCENTILE_BREAKS)
}
\arguments{
\item{.data}{a dplyr::tbl}

\item{breaks}{named numeric vector}
}
\description{
To compute CES2 scores
}

