% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute_CES2_subscores}
\alias{compute_CES2_subscores}
\title{To compute CES2 subscores (requires pre-calculated Pctl column)}
\usage{
compute_CES2_subscores(.data, min_obs = 4, ...)
}
\arguments{
\item{.data}{a dplyr::tbl}

\item{min_obs}{minimum number of observations in a group (else NA)}

\item{\dots}{ignored}
}
\description{
To compute CES2 subscores (requires pre-calculated Pctl column)
}

