% Generated by roxygen2 (4.0.1): do not edit by hand
\name{IDs}
\alias{IDs}
\alias{IDs.SpatialPolygons}
\alias{IDs.SpatialPolygonsDataFrame}
\title{To get the IDs of features in a SpatialPolygons* object}
\usage{
IDs(spobj, ...)

\method{IDs}{SpatialPolygons}(spobj, ...)

\method{IDs}{SpatialPolygonsDataFrame}(spobj, ...)
}
\arguments{
\item{spobj}{SpatialPolygons* object}

\item{\dots}{ignored}
}
\description{
To get the IDs of features in a SpatialPolygons* object
}

