% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.SpatialPolygonsDataFrame}
\alias{merge.SpatialPolygonsDataFrame}
\title{Merge a SpatialPolygonsDataFrame with a data.frame}
\usage{
\method{merge}{SpatialPolygonsDataFrame}(x, y, by, ...)
}
\arguments{
\item{x}{SpatialPolygonsDataFrame}

\item{y}{data.frame}

\item{by}{character}

\item{\dots}{ignored}
}
\description{
Merge a SpatialPolygonsDataFrame with a data.frame
}

