% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pctl}
\alias{pctl}
\title{Calculate percentiles (as in CalEnviroScreen 2.0)}
\usage{
pctl(x, ...)
}
\arguments{
\item{x}{values}

\item{\dots}{futher arguments to \link{comp_rank}}
}
\description{
Calculate percentiles (as in CalEnviroScreen 2.0)
}

