% Generated by roxygen2 (4.0.1): do not edit by hand
\name{percentiles}
\alias{percentiles}
\title{Calculate percentiles, as in CalEnviroScreen 2.0}
\usage{
percentiles(x, ...)
}
\arguments{
\item{x}{values}

\item{\dots}{futher arguments to \link{ranks}}
}
\description{
Calculate percentiles, as in CalEnviroScreen 2.0
}

