% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_kmz}
\alias{read_kmz}
\title{To unzip and import the "doc.kml" portion of a KMZ file}
\usage{
read_kmz(file, layer)
}
\arguments{
\item{file}{filename}

\item{layer}{layer name}
}
\description{
To unzip and import the "doc.kml" portion of a KMZ file
}

