% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reindex}
\alias{reindex}
\alias{reindex.SpatialPolygonsDataFrame}
\alias{reindex.data.frame}
\title{To reindex a data.frame or Spatial*DataFrame (effectively replacing row.names)}
\usage{
reindex(x, replace)

\method{reindex}{data.frame}(x, replace)

\method{reindex}{SpatialPolygonsDataFrame}(x, replace)
}
\arguments{
\item{x}{data.frame or Spatial*DataFrame}

\item{replace}{named character vector}
}
\description{
To reindex a data.frame or Spatial*DataFrame (effectively replacing row.names)
}

