% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weighted_average}
\alias{weighted_average}
\title{To compute a weighted average}
\usage{
weighted_average(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{w}{weights}

\item{na.rm}{logical}
}
\description{
To compute a weighted average
}

