% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CA_tracts}
\alias{CA_tracts}
\title{California census tracts}
\format{SpatialPolygonsDataFrame}
\source{
\url{http://oehha.ca.gov/ej/ces2.html}
}
\usage{
CA_tracts
}
\description{
California census tracts, with areas computed from an Albers projection
}

