% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CES2_percentile_breaks}
\alias{CES2_percentile_breaks}
\title{CalEnviroScreen 2.0 percentile breaks}
\format{named numeric vector}
\usage{
CES2_percentile_breaks
}
\description{
Percentile breaks inferred from published CES 2.0 data
}

