% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalEnviroScreen-data.R
\docType{data}
\name{CES2_PERCENTILE_BREAKS}
\alias{CES2_PERCENTILE_BREAKS}
\title{Percentile breaks (inferred from published CES 2.0 data)}
\format{named numeric vector}
\usage{
data(CES2_metadata); CES2_PERCENTILE_BREAKS
}
\description{
Percentile breaks (inferred from published CES 2.0 data)
}

