% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalEnviroScreen-data.R
\docType{data}
\name{CES2_POPCHAR_VARS}
\alias{CES2_POPCHAR_VARS}
\title{Variables associated with population characteristics}
\format{named character vector, with names corresponding to column names in the published Excel file}
\usage{
data(CES2_metadata); CES2_POPCHAR_VARS
}
\description{
Names of variables used to calculate population characteristics score
}

