% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalEnviroScreen-data.R
\docType{data}
\name{CES2_scores}
\alias{CES2_scores}
\title{CalEnviroScreen 2.0 scores}
\format{A data.frame with 8035 rows and 5 columns}
\source{
\url{http://oehha.ca.gov/ej/ces2.html}
}
\usage{
data(CES2); CES2_scores
}
\description{
CalEnviroScreen 2.0 scores (as published)
}

