% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{CES2_WEIGHTS}
\alias{CES2_WEIGHTS}
\title{Relative weights used to compute CES 2.0 scores}
\format{named numeric vector}
\usage{
data(CES2_metadata); CES2_WEIGHTS
}
\description{
Relative weights used to compute CES 2.0 scores
}

